pyproject.toml
[tool.poetry]
name = 'bot'
version = '1.0.0'
description = 'Simple starter template for a discord.py project.'
authors = ['Rob Wagner <rob@robwagner.dev>']
license = ''
readme = 'README.md'
homepage = 'https://github.com/robertwayne/dpy-cogwatch-poetry-template'
repository = 'https://github.com/robertwayne/dpy-cogwatch-poetry-template'

[tool.poetry.scripts]
bot = 'bot.client:__poetry_run'
fmt = 'scripts:fmt'

[tool.poetry.dependencies]
python = '^3.9'
'discord.py' = '^1.5'
cogwatch = '^2.1'
uvloop = '^0.15'

[tool.poetry.dev-dependencies]
black = '^20.8b1'
python-dotenv = '^0.15'

[tool.black]
line-length = 120
skip-string-normalization = 1
target-version = ['py37', 'py38']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | docs
    | assets
    | venv
  )/
)
'''

[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'
